attempting post reservation with seat status

+++ omas/src/main/java/com/omasystem/omas/Controller/ReservationRestController.java
@@ -3,6 +3,7 @@ package com.omasystem.omas.Controller;
 import java.util.Map;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -30,7 +31,7 @@ public class ReservationRestController {
     }
 
     @PostMapping("/{seat_id}/add")
-    public Map<String, Object> insertReservation(@PathVariable Long seat_id, @RequestBody ReservationInputBodyModel body)
+    public Map<String, Object> insertReservation(@PathVariable int seat_id, @RequestBody ReservationInputBodyModel body)
     {
         return reservationService.insertReservation(seat_id, body);
     }
